requirements.txt
python-dotenv
pycryptodome
pyyaml
